+++ app/src/main/java/com/kiero/data/parent/mission/di/MissionRepositoryModule.kt
@@ -1,11 +1,11 @@
-package com.kiero.data.mission.di
+package com.kiero.data.parent.mission.di
 
-import com.kiero.data.mission.repository.AutoMissionRepository
-import com.kiero.data.mission.repository.MissionRepository
-import com.kiero.data.mission.repositoryimpl.AutoMissionRepositoryImpl
-import com.kiero.data.mission.repositoryimpl.MissionRepositoryImpl
-import com.kiero.data.parent.repository.ParentMissionAddRepository
-import com.kiero.data.parent.repositoryimpl.ParentMissionAddRepositoryImpl
+import com.kiero.data.parent.mission.repositoryimpl.AutoMissionRepositoryImpl
+import com.kiero.data.parent.mission.repositoryimpl.MissionRepositoryImpl
+import com.kiero.domain.repository.parent.mission.AutoMissionRepository
+import com.kiero.domain.repository.parent.mission.MissionRepository
+import com.kiero.domain.repository.parent.mission.ParentMissionAddRepository
+import com.kiero.data.parent.mission.repositoryimpl.ParentMissionAddRepositoryImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
