+++ app/src/main/java/com/kiero/data/parent/mission/remote/dto/response/MissionCompleteResponseDto.kt
@@ -1,14 +1,12 @@
-package com.kiero.data.mission.remote.dto.response
+package com.kiero.data.parent.mission.remote.dto.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
-typealias MissionBulkCreateResponseDto = List<CreatedMissionDto>
-
 @Serializable
-data class CreatedMissionDto(
+data class MissionCompleteResponseDto(
     @SerialName("id")
-    val id: Long,
+    val id : Long,
     @SerialName("name")
     val name: String,
     @SerialName("reward")
